rename "Abstractions" to "Providers"
