+++ src/lib/Profiles/Core/Actions/Log.ts
@@ -1,6 +1,6 @@
-import ILogger from '../../../Abstractions/ILogger';
 import Node from '../../../Nodes/Node';
 import NodeEvalContext from '../../../Nodes/NodeEvalContext';
+import ILogger from '../../../Providers/ILogger';
 import Socket from '../../../Sockets/Socket';
 
 export default class Log extends Node {
