+++ src/lib/Profiles/Core/Events/OnLifecycleTick.ts
@@ -1,6 +1,6 @@
-import ILifecycleEventEmitter from '../../../Abstractions/ILifecycleEventEmitter';
 import Node from '../../../Nodes/Node';
 import NodeEvalContext from '../../../Nodes/NodeEvalContext';
+import ILifecycleEventEmitter from '../../../Providers/ILifecycleEventEmitter';
 import Socket from '../../../Sockets/Socket';
 
 // inspired by: https://docs.unrealengine.com/4.27/en-US/ProgrammingAndScripting/Blueprints/UserGuide/Events/
