+++ src/lib/Profiles/SceneGraph/Queries/GetSceneNodeProperty.ts
@@ -1,7 +1,7 @@
 import { Object3D } from 'three';
 
-import IThree from '../../../Abstractions/IThree';
 import Node from '../../../Nodes/Node';
+import IThree from '../../../Providers/IThree';
 import Socket from '../../../Sockets/Socket';
 
 export default class GetSceneNodeProperty<T> extends Node {
