+++ src/lib/Providers/ILifecycleEventEmitter.ts
