+++ src/lib/Providers/IThree.ts
@@ -1,5 +1,6 @@
+/* eslint-disable semi */
 import { Object3D } from 'three';
 
 export default interface IThree {
     getObject3D(nodeId: string): Object3D;
-};;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+};
