+++ src/lib/Providers/Implementations/ManualLifecycleEventEmitter.ts
