+++ src/lib/index.ts
@@ -36,11 +36,11 @@ export { default as ValueTypeRegistry } from './Values/ValueTypeRegistry';
 export { default as Registry } from './Registry';
 
 // interface registry
-export { default as ImplementationRegistry } from './Abstractions/ImplementationRegistry';
-export { default as ILifecycleEventEmitter } from './Abstractions/ILifecycleEventEmitter';
-export { default as DefaultLogger } from './Abstractions/Implementations/DefaultLogger';
-export { default as ManualLifecycleEventEmitter } from './Abstractions/Implementations/ManualLifecycleEventEmitter';
-export { default as ILogger } from './Abstractions/ILogger';
+export { default as ImplementationRegistry } from './Providers/ImplementationRegistry';
+export { default as ILifecycleEventEmitter } from './Providers/ILifecycleEventEmitter';
+export { default as DefaultLogger } from './Providers/Implementations/DefaultLogger';
+export { default as ManualLifecycleEventEmitter } from './Providers/Implementations/ManualLifecycleEventEmitter';
+export { default as ILogger } from './Providers/ILogger';
 
 export { default as registerCoreProfile } from './Profiles/Core/registerCoreProfile';
 export { default as registerSceneGraphProfile } from './Profiles/SceneGraph/registerSceneGraphProfile';
